fix(qa): rendre Psalm vert — opérandes int/float explicites
strictBinaryOperands est actif par défaut chez Psalm 6 : les mélanges
int/float des calculs de durées étaient signalés en InvalidOperand. Cast
explicite sur les cinq sites plutôt que désactivation du contrôle, qui
relâcherait aussi la vérification des concaténations.

Le MapField de recordCancellationDelivered est suppressé au site : les
stubs google/protobuf typent les constantes GPBType en `long`, comme
déjà admis en baseline pour l'autre construction du même fichier.

+++ src/Durable/Activity/ActivityOptions.php
@@ -80,7 +80,7 @@ final readonly class ActivityOptions
         }
 
         return $this->initialInterval
-            ->multipliedBy($this->backoffCoefficient ** ($nextAttempt - 2))
+            ->multipliedBy($this->backoffCoefficient ** (float) ($nextAttempt - 2))
             ->shortest($this->effectiveMaximumInterval());
     }
